✨ Python 3.12 support
This commit adds Python 3.12 in the list of programming languages in the pyproject.toml file and also in the noxfile.py for testing purposes.
Due to qiskit-terra not supporting Python 3.12 yet, tests for this version are skipped for now.
Lastly, the version of cibuildwheel used in the GitHub Actions workflow was updated to v2.15 for building wheels.

Signed-off-by: burgholzer <burgholzer@me.com>