✨ Adopt `uv` in CI (#368)
## Description

This PR updates the CI configuration to use uv by Astral (the creators
of ruff), which is an extremely fast Python package installer and
resolver, written in Rust. It is sesigned as a drop-in replacement for
common pip and pip-tools workflows.
See https://github.com/astral-sh/uv

Similar to ruff, it is extremely fast and in very active development.
This should speed up any of the Python-based CI runs without much
confirguration overhead.

This PR refactors the `minimums` check to make use of `uv`'s
`--resolution=lowest-direct" feature. This allows us to completely get
rid of the `constraints.txt` file.
Furthermore, it makes use of the new versions of `mqt-core`'s reusable
workflows.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>